More tests and refactoring
