+++ metrics/core/lib/src/domain/value_objects/exceptions/validation_exception.dart
