+++ metrics/core/test/domain/value_objects/exceptions/email_validation_exception_test.dart
@@ -6,5 +6,13 @@ void main() {
     test("throws an ArgumentError if the given code is null", () {
       expect(() => EmailValidationException(null), throwsArgumentError);
     });
+
+    test("successfully creates with the given error code", () {
+      const errorCode = EmailValidationErrorCode.isNull;
+
+      final validationException = EmailValidationException(errorCode);
+
+      expect(validationException.code, equals(errorCode));
+    });
   });
 }
