+++ metrics/core/test/domain/value_objects/exceptions/password_validation_exception_test.dart
@@ -6,5 +6,13 @@ void main() {
     test("throws an ArgumentError if the given code is null", () {
       expect(() => PasswordValidationException(null), throwsArgumentError);
     });
+
+    test("successfully creates with the given error code", () {
+      const errorCode = PasswordValidationErrorCode.isNull;
+
+      final validationException = PasswordValidationException(errorCode);
+
+      expect(validationException.code, equals(errorCode));
+    });
   });
 }
